pyproject.toml
[project]
name = "cubeclient"

[tool.isort]
profile = "black"
lines_after_imports = 2
honor_noqa = true

[tool.black]
line-length = 119

[tool.poetry]
name = "cubeclient"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
Pillow = "~9.4"
antlr4-python3-runtime = "~4.9"
appdirs = "~1.4"
ijson = "~3.2"
immutabledict = "~3.0"
lazy-property = "~0.0"
more-itertools = "~10.1"
multiset = "~3.0"
numpy = "~1.26"
pickledb = "0.9"
pillow = "~9.4"
promise = "~2.3"
requests = "^2.31.0"
sqlalchemy = "~2.0"
bidict = "^0.23.0"
cachetools = "^5.3.2"
simplejson = "~3.19"
websocket = "~0.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
